Moved the encoding data classes
